requirements.txt
nionswift-instrumentation==0.20.8
pyserial>=3.5
requests>=2.27.1